Refactor frontend component file structure
