+++ frontend/src/components/chat/BotMessage.js
