+++ frontend/src/components/sidebar/HiddenSidebar.js
